+++ tests/testthat/test-qual_T7_2.R
@@ -32,12 +32,7 @@ testthat::test_that("Qual: Protocol Deviation Assessments can be done correctly 
   iwalk(
     test,
     ~ expect_equal(
-      n_distinct(.x$Mapped_SUBJ[[
-        kri_workflows[[.y]]$steps[[which(
-          map_chr(kri_workflows[[.y]]$steps, ~ .x$name) ==
-            "gsm.core::Input_Rate"
-        )]]$params$strGroupCol
-      ]]),
+      CountTransformedGroups(.x$Analysis_Input),
       nrow(.x$Analysis_Transformed)
     )
   )
@@ -61,7 +56,7 @@ testthat::test_that("Qual: Protocol Deviation Assessments can be done correctly 
   iwalk(
     test_custom,
     ~ expect_equal(
-      n_distinct(.x$Mapped_SUBJ[[tolower(kri_custom[[.y]]$meta$GroupLevel)]]),
+      CountTransformedGroups(.x$Analysis_Input),
       nrow(.x$Analysis_Transformed)
     )
   )